requirements.txt
asgiref==3.2.10
black
Django==3.1.1
django-cors-headers==3.5.0
django-debug-toolbar
django-templated-mail==1.1.1
django-cleanup
django-ical
django-rest-registration
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.1
drf-yasg
drf-yasg[validation]
flake8
isort
gunicorn
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
pytest-cov
pytest-django
pytest-xdist
pytest
requests
six
sqlparse==0.3.1
xlwt